Add public cosign key
We are starting to use cosign to sign our container
images to increase supply chain security.

To make verification of this signature as easy as
possible, we need to make the public key accessible
on a source that users trust, e.g. a companies
homepage.

In the future, a signature can be verified like this

cosign verify qdrant/qdrant:v1.6.1 --key https://qdrant.tech/cosign.pub